# openapi.yml
openapi: 3.0.3
info:
  title: Mini WMS - API
  description: API REST para microaplicação Mini WMS (controle de estoque, lotes, ordens e movimentações).
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
  /produtos:
    get:
      summary: Listar produtos
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
    post:
      summary: Criar produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoInput'
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
  /produtos/{produtoId}:
    parameters:
      - name: produtoId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Obter produto por id
      responses:
        '200':
          description: Produto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          description: Não encontrado
    put:
      summary: Atualizar produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoInput'
      responses:
        '200':
          description: Produto atualizado
    delete:
      summary: Excluir (lógico) produto
      responses:
        '204':
          description: Excluído (inativo)

  /enderecos:
    get:
      summary: Listar endereços
      responses:
        '200':
          description: Lista de endereços
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Endereco'
    post:
      summary: Criar endereço
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnderecoInput'
      responses:
        '201':
          description: Endereço criado

  /lotes:
    get:
      summary: Listar lotes
      responses:
        '200':
          description: Lista de lotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lote'
    post:
      summary: Criar lote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoteInput'
      responses:
        '201':
          description: Lote criado

  /ordens:
    get:
      summary: Listar ordens de saída
      responses:
        '200':
          description: Lista de ordens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrdemSaida'
    post:
      summary: Criar ordem de saída (realiza reserva)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdemSaidaInput'
      responses:
        '201':
          description: Ordem criada e estoque reservado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdemSaida'
  /ordens/{ordemId}:
    parameters:
      - name: ordemId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Obter ordem por id
      responses:
        '200':
          description: Ordem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdemSaida'
  /ordens/{ordemId}/confirmar:
    post:
      summary: Confirmar processamento da ordem (gera movimentações)
      parameters:
        - name: ordemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ordem processada
  /movimentacoes:
    get:
      summary: Listar movimentações
      responses:
        '200':
          description: Lista de movimentações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movimentacao'
    post:
      summary: Registrar movimentação (entrada/transferência/saída)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimentacaoInput'
      responses:
        '201':
          description: Movimentação registrada

  /usuarios:
    get:
      summary: Listar usuários
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
    post:
      summary: Criar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '201':
          description: Usuário criado

components:
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        email:
          type: string
          format: email
        perfil:
          type: string
        ativo:
          type: boolean
        criado_em:
          type: string
          format: date-time
      required: [id, nome, email]

    UsuarioInput:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        senha:
          type: string
        perfil:
          type: string
      required: [nome, email, senha]

    Produto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        nome:
          type: string
        descricao:
          type: string
        unidade_medida:
          type: string
        peso:
          type: number
        ativo:
          type: boolean
      required: [id, sku, nome]

    ProdutoInput:
      type: object
      properties:
        sku:
          type: string
        nome:
          type: string
        descricao:
          type: string
        unidade_medida:
          type: string
        peso:
          type: number
      required: [sku, nome]

    Endereco:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codigo:
          type: string
        descricao:
          type: string
        tipo:
          type: string
        capacidade:
          type: integer
        ativo:
          type: boolean
      required: [id, codigo]

    EnderecoInput:
      type: object
      properties:
        codigo:
          type: string
        descricao:
          type: string
        tipo:
          type: string
        capacidade:
          type: integer
      required: [codigo]

    Lote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        produto_id:
          type: string
        endereco_id:
          type: string
        codigo_lote:
          type: string
        quantidade_total:
          type: number
        quantidade_reservada:
          type: number
        data_fabricacao:
          type: string
          format: date
        data_validade:
          type: string
          format: date
        status:
          type: string
      required: [id, produto_id, codigo_lote, quantidade_total]

    LoteInput:
      type: object
      properties:
        produto_id:
          type: string
        endereco_id:
          type: string
        codigo_lote:
          type: string
        quantidade_total:
          type: number
        data_fabricacao:
          type: string
          format: date
        data_validade:
          type: string
          format: date
      required: [produto_id, codigo_lote, quantidade_total]

    ItemOrdemSaida:
      type: object
      properties:
        id:
          type: string
        ordem_saida_id:
          type: string
        produto_id:
          type: string
        lote_id:
          type: string
        quantidade:
          type: number
        unidade:
          type: string
        status_item:
          type: string

    OrdemSaida:
      type: object
      properties:
        id:
          type: string
        codigo:
          type: string
        data_criacao:
          type: string
          format: date-time
        data_confirmacao:
          type: string
          format: date-time
        cliente:
          type: string
        status:
          type: string
        usuario_id:
          type: string
        itens:
          type: array
          items:
            $ref: '#/components/schemas/ItemOrdemSaida'
      required: [id, codigo, status]

    OrdemSaidaInput:
      type: object
      properties:
        cliente:
          type: string
        usuario_id:
          type: string
        itens:
          type: array
          items:
            type: object
            properties:
              produto_id:
                type: string
              quantidade:
                type: number
            required: [produto_id, quantidade]
      required: [itens, usuario_id]

    Movimentacao:
      type: object
      properties:
        id:
          type: string
        tipo:
          type: string
        produto_id:
          type: string
        lote_id:
          type: string
        origem_endereco_id:
          type: string
        destino_endereco_id:
          type: string
        quantidade:
          type: number
        data_movimentacao:
          type: string
          format: date-time
        usuario_id:
          type: string
        observacao:
          type: string

    MovimentacaoInput:
      type: object
      properties:
        tipo:
          type: string
        produto_id:
          type: string
        lote_id:
          type: string
        origem_endereco_id:
          type: string
        destino_endereco_id:
          type: string
        quantidade:
          type: number
        usuario_id:
          type: string
        observacao:
          type: string
      required: [tipo, produto_id, quantidade, usuario_id]

  responses:
    NotFound:
      description: Recurso não encontrado
    BadRequest:
      description: Requisição inválida